export db feat fixed rotation game borded gone
