+++ app/src/main/res/layout-large/activity_rotation_game.xml
@@ -84,7 +84,6 @@
             android:id="@+id/RG_rectToMatch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:background="@drawable/border"
             android:textSize="80sp" />
