+++ app/src/main/res/layout-xlarge/activity_rotation_game.xml
@@ -81,7 +81,6 @@
             android:id="@+id/RG_rectToMatch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:background="@drawable/border"
             android:textSize="100sp" />
