+++ app/src/main/res/layout/activity_rotation_game.xml
@@ -79,7 +79,6 @@
             android:id="@+id/RG_rectToMatch"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:background="@drawable/border"
             android:textSize="60sp" />
